UserController & AdminController implemented and tested

+++ src/main/java/com/agromov/votemeal/util/LunchBuilder.java
@@ -41,7 +41,7 @@ public class LunchBuilder
         return this;
     }
 
-    public LunchBuilder withPrice(int price)
+    public LunchBuilder withPrice(float price)
     {
         lunch.setPrice(price);
         return this;
